feat(lab): add lab 3 for mock out external dep
