+++ sessions/mock_out_external_dep/lab_03_risk/foo.py
@@ -1,0 +1,6 @@
+def add_one(num):
+    return num + 1
+
+
+def add_one_and_mul_two(num):
+    return add_one(num) * 2
